Extract handler into separate file
