+++ projects/archiver/src/tasks/issue/index.ts
@@ -6,7 +6,7 @@ import { Bucket } from '../../utils/s3'
 import { getPublishedId } from '../../utils/path-builder'
 import { putStatus } from '../../services/status'
 import { logInput, logOutput } from '../../utils/log'
-import { handleAndNotify } from '../../services/pub-status-notifier'
+import { handleAndNotify } from '../../services/task-handler'
 
 export interface IssueParams {
     issuePublication: IssuePublicationIdentifier
